test: consolidate coverage tests Signed-off-by: Gregor Zeitlinger <gregor.zeitlinger@grafana.com>
