--- assert/assert_more_test.go
@@ -1,10 +1,0 @@
-package assert
-
-import "testing"
-
-func TestFailureError(t *testing.T) {
-	got := (Failure{Rule: "contains", Detail: "missing"}).Error()
-	if got != "contains: missing" {
-		t.Fatalf("Failure.Error() = %q", got)
-	}
-}
