--- report/report_more_test.go
@@ -1,12 +1,0 @@
-package report
-
-import "testing"
-
-func TestReportersClose(t *testing.T) {
-	if err := NewTextReporter(nil, VerboseDefault).Close(); err != nil {
-		t.Fatalf("TextReporter.Close: %v", err)
-	}
-	if err := NewNDJSONReporter(nil, VerboseDefault).Close(); err != nil {
-		t.Fatalf("NDJSONReporter.Close: %v", err)
-	}
-}
