+++ report/report_test.go
@@ -264,3 +264,12 @@ func TestSplitSource(t *testing.T) {
 		}
 	}
 }
+
+func TestReportersClose(t *testing.T) {
+	if err := NewTextReporter(nil, VerboseDefault).Close(); err != nil {
+		t.Fatalf("TextReporter.Close: %v", err)
+	}
+	if err := NewNDJSONReporter(nil, VerboseDefault).Close(); err != nil {
+		t.Fatalf("NDJSONReporter.Close: %v", err)
+	}
+}
